% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractClass.R
\docType{class}
\name{AbstractClass-class}
\alias{AbstractClass-class}
\alias{AbstractClass}
\title{Reference Class representing a non instantiable class}
\description{
An abstract base class with some utility methods
}

