% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{as.character,Attribute-method}
\alias{as.character,Attribute-method}
\title{as.character for Attribute classes}
\usage{
\S4method{as.character}{Attribute}(x)
}
\arguments{
\item{x}{the object to convert}
}
\description{
as.character for Attribute classes
}
