% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Element.R
\name{as.character,Element-method}
\alias{as.character,Element-method}
\title{as.character for Element classes}
\usage{
\S4method{as.character}{Element}(x)
}
\arguments{
\item{x}{the object to convert}
}
\description{
as.character for Element classes
}
