% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Namespace.R
\name{as.character,Namespace-method}
\alias{as.character,Namespace-method}
\title{as.character for Namespace classes}
\usage{
\S4method{as.character}{Namespace}(x)
}
\arguments{
\item{x}{the object to convert}
}
\description{
as.character for Namespace classes
}
