% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Text.R
\name{as.character,Text-method}
\alias{as.character,Text-method}
\title{as.character for Text classes}
\usage{
\S4method{as.character}{Text}(x)
}
\arguments{
\item{x}{the object to convert}
}
\description{
as.character for Text classes
}
