% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{as.vector,Attribute-method}
\alias{as.vector,Attribute-method}
\title{as.vector for Attribute classes}
\usage{
\S4method{as.vector}{Attribute}(x)
}
\arguments{
\item{x}{the object to convert}
}
\description{
as.vector for Attribute classes
}
