% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Element.R
\name{as.vector,Element-method}
\alias{as.vector,Element-method}
\title{as.vector for Element classes}
\usage{
\S4method{as.vector}{Element}(x)
}
\arguments{
\item{x}{the object to convert}
}
\description{
as.vector for Element classes
}
