% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Namespace.R
\name{as.vector,Namespace-method}
\alias{as.vector,Namespace-method}
\title{as.vector for Namespace classes}
\usage{
\S4method{as.vector}{Namespace}(x)
}
\arguments{
\item{x}{the object to convert}
}
\description{
as.vector for Namespace classes
}
