% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Text.R
\name{as.vector,Text-method}
\alias{as.vector,Text-method}
\title{as.vector for Text classes}
\usage{
\S4method{as.vector}{Text}(x)
}
\arguments{
\item{x}{the object to convert}
}
\description{
as.vector for Text classes
}
