% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\docType{class}
\name{Attribute-class}
\alias{Attribute-class}
\alias{Attribute}
\alias{as.vector,Attribute-method}
\alias{as.character,Attribute-method}
\title{A reference class representing an tag (element) attribute}
\usage{
\S4method{as.vector}{Attribute}(x)

\S4method{as.character}{Attribute}(x)
}
\arguments{
\item{x}{the object to convert}

\item{name}{The name (character) of the new Attribute}

\item{value}{The value (character) of the new Attribute}
}
\description{
A reference class representing an tag (element) attribute

as.vector for Attribute classes

as.character for Attribute classes
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.vector}: as.vector(Attribute)

\item \code{as.character}: as.character(Attribute)
}}

\section{Fields}{

\describe{
\item{\code{name}}{The local name of the Attribute}

\item{\code{value}}{The value (character) of the Attribute}
}}


