% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Content.R
\docType{class}
\name{Content-class}
\alias{Content-class}
\alias{Content}
\title{A reference class representing content that can belong to an Element}
\description{
A reference class representing content that can belong to an Element
}

