% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document.R
\docType{class}
\name{Document-class}
\alias{Document-class}
\alias{Document}
\title{Reference Class representing an XML document}
\description{
The base container for the DOM
}
\details{
Methods allow access to the root element as well as the
DocType and other document-level information.
}
\section{Fields}{

\describe{
\item{\code{content}}{This document's content including comments, PIs, a possible DocType, and a root element.}

\item{\code{baseURI}}{see https://www.w3.org/TR/2003/WD-DOM-Level-3-Core-20030226/core.html#baseURIs-Considerations}
}}


