% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomBuilder.R
\docType{class}
\name{DomBuilder-class}
\alias{DomBuilder-class}
\alias{DomBuilder}
\title{Create a rcdom object tree based on parsing events}
\description{
Create a rcdom object tree based on parsing events
}
\section{Methods}{

\describe{
\item{\code{endDocument()}}{Event signalling parsing has completed}

\item{\code{endElement(name)}}{end element event; @param name the element name}

\item{\code{startDocument()}}{Event signalling parsing has begun}

\item{\code{startElement(name, attributes)}}{start element event; @param name the element name, @param attributes a named list of attributes}

\item{\code{text(text)}}{text event; @param text the character content of the Text node}
}}

