% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Element.R
\docType{class}
\name{Element-class}
\alias{Element-class}
\alias{Element}
\title{Reference class representing an XML tag}
\arguments{
\item{name}{The name of the tag (optional)}

\item{namespace}{a namespace object setting the context for the the element (optional)}
}
\description{
Reference class representing an XML tag
}
\details{
An XML element. Methods allow the user to get and manipulate its child
elements and content, directly access the element's textual content,
manipulate its attributes, and manage namespaces.
}
\section{Fields}{

\describe{
\item{\code{name}}{The local name of the element}

\item{\code{namespace}}{The namespace of the element}

\item{\code{parent}}{The document if this is the root element, otherwise the parent element}

\item{\code{attributeList}}{a list of all the attributes belonging to this element}

\item{\code{contentList}}{all the children of this element}
}}


