% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parent.R
\docType{class}
\name{Parent-class}
\alias{Parent-class}
\alias{Parent}
\title{An abstract class for rDOM objects which are allowed to contain Content, Element Document.
Not sure what to do with
getDescendants()
canContainContent()
...omitting them for now}
\description{
An abstract class for rDOM objects which are allowed to contain Content, Element Document.
Not sure what to do with
getDescendants()
canContainContent()
...omitting them for now
}

