% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser.R
\docType{class}
\name{Parser-class}
\alias{Parser-class}
\alias{Parser}
\title{Parse an xml string and create sax like events}
\description{
an XML parser based on an article on creating a quick and dirty xml parser by Steven Brandt:
https://www.javaworld.com/article/2077493/java-tip-128--create-a-quick-and-dirty-xml-parser.html
}

