% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stack.R
\docType{class}
\name{Stack-class}
\alias{Stack-class}
\alias{Stack}
\title{A general purpose linked stack}
\description{
A general purpose linked stack
}
\section{Fields}{

\describe{
\item{\code{size}}{the size of the stack (number of elements in the stack)}

\item{\code{stackNode}}{an envronment containing the current element and the one under}
}}

\section{Methods}{

\describe{
\item{\code{peek()}}{Get the top element from the stack without changing it}

\item{\code{pop()}}{Pull the top element from the stack removing it from the stack}

\item{\code{push(val)}}{Add an element to the top of the stack}

\item{\code{size()}}{Get the current size of the stack}
}}

