% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isRc}
\alias{isRc}
\title{Common utility functions}
\usage{
isRc(x, clazz = "refClass")
}
\arguments{
\item{x}{the object to check}

\item{clazz}{the name of the class e.g. "Element" for the Element class.
Optional, if omitted it checks that the object is a reference class}
}
\value{
A boolean indicating whether the object \code{x} belongs to the class or not
}
\description{
Common utility functions
}
\section{Functions}{
\itemize{
\item \code{isRc}: Check if the object is a reference class, similar to isS4().
}}

