% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlConverter.R
\name{xmlConverter}
\alias{xmlConverter}
\alias{xmlrToDataFrame}
\title{XML conversion functions}
\usage{
xmlrToDataFrame(element)
}
\arguments{
\item{element}{the element to convert}
}
\value{
a data frame
}
\description{
this is a convenience method to take all the children of the given Element
and create a data frame based on the content of each child where each child constitutes a row
and the attributes or elements (including text) will constiture the columns.
It assumes a homeogenous structure and the column names are takes from the first child
}
\section{Functions}{
\itemize{
\item \code{xmlrToDataFrame}: create a data frame from a xmlr Element
}}

