% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XmlImporter.R
\name{XmlImporter}
\alias{XmlImporter}
\alias{parse.xmlstring}
\alias{parse.xmlfile}
\title{XML import functions}
\usage{
parse.xmlstring(xml)

parse.xmlfile(fileName)
}
\arguments{
\item{xml}{an xml character string to parse}

\item{fileName}{the name of the xml file to parse}
}
\value{
a Document object

a Document object
}
\description{
XML import functions
}
\section{Functions}{
\itemize{
\item \code{parse.xmlstring}: create a Document from a character string

\item \code{parse.xmlfile}: create a Document from a xml file
}}

