% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlConverter.R
\name{xmlrToDataFrame}
\alias{xmlrToDataFrame}
\title{Create a data frame from a xmlr Element}
\usage{
xmlrToDataFrame(element)
}
\arguments{
\item{element}{the element to convert}
}
\value{
a data frame
}
\description{
This is a convenience method to take all the children of the given Element
and create a data frame based on the content of each child where each child constitutes a row
and the attributes or elements (including text) will constitute the columns.
It assumes a homogeneous structure and the column names are takes from the first child
}
