% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\docType{class}
\name{Attribute-class}
\alias{Attribute-class}
\alias{Attribute}
\title{A reference class representing an tag (element) attribute}
\description{
A reference class representing an tag (element) attribute
}

