% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Content.R
\docType{class}
\name{Content-class}
\alias{Content-class}
\alias{Content}
\title{An abstract reference class representing content that can belong to an Element}
\description{
An abstract reference class representing content that can belong to an Element
}
\section{Fields}{

\describe{
\item{\code{m_parent}}{the parent (if any)}
}}


