% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document.R
\docType{class}
\name{Document-class}
\alias{Document-class}
\alias{Document}
\alias{as.vector,Document-method}
\alias{as.character,Document-method}
\title{Reference Class representing an XML document}
\usage{
\S4method{as.vector}{Document}(x)

\S4method{as.character}{Document}(x)
}
\arguments{
\item{x}{the object to convert}
}
\description{
The base container for the DOM
}
\details{
Methods allow access to the root element as well as the
DocType and other document-level information.
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.vector}: as.vector(Document)

\item \code{as.character}: as.character(Document)
}}


