% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Element.R
\docType{class}
\name{Element-class}
\alias{Element-class}
\alias{Element}
\alias{as.vector,Element-method}
\alias{as.character,Element-method}
\title{Reference class representing an XML tag}
\usage{
\S4method{as.vector}{Element}(x)

\S4method{as.character}{Element}(x)
}
\arguments{
\item{x}{the object to convert}

\item{...}{The name of the tag (optional) and namespace a namespace object setting the context for the the element (optional)}
}
\description{
Reference class representing an XML tag

as.vector for Element classes

as.character for Element classes
}
\details{
An XML element. Methods allow the user to get and manipulate its child
elements and content, directly access the element's textual content,
manipulate its attributes, and manage namespaces.
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.vector}: as.vector(Element)

\item \code{as.character}: as.character(Element)
}}

\section{Fields}{

\describe{
\item{\code{name}}{The local name of the element}

\item{\code{namespace}}{The namespace of the element}

\item{\code{parent}}{The document if this is the root element, otherwise the parent element}

\item{\code{attributeList}}{a list of all the attributes belonging to this element}

\item{\code{contentList}}{all the children of this element}
}}


