% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Namespace.R
\docType{class}
\name{Namespace-class}
\alias{Namespace-class}
\alias{Namespace}
\title{A reference class representing the context (Namespace)}
\description{
A reference class representing the context (Namespace)
}

