% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Text.R
\docType{class}
\name{Text-class}
\alias{Text-class}
\alias{Text}
\title{Reference class representing text content}
\description{
Reference class representing text content
}
\details{
An XML character sequence. Provides a modular, parentable method of representing text.
}

