% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printp}
\alias{printp}
\alias{printp0}
\alias{isRc}
\title{Common utility functions}
\usage{
printp(...)

printp0(...)

isRc(x)
}
\arguments{
\item{...}{all the stuff that should be printed}

\item{x}{the object to check}
}
\description{
Common utility functions
}
\section{Functions}{
\itemize{
\item \code{printp}: Simple combo of print and paste.

\item \code{printp0}: Simple combo of print and paste0.

\item \code{isRc}: Check if the object is a reference class, similar to isS4().
}}

